Update tag.py
tag.state is always a dict (could be None before)

But only the "main/managed tag" saves its state in the session dict. Others use an internal dict to manage the state

Tag.state is now the better place to store internal values in the tag (avoid cluttering with self reference). And as it's not serialized in the session, all kind of types can be stored (for other tags ! the main one no!)

UT should be broken after this commit ;-)